requirements.txt
kafka-python
kafka-python-ng
faker